Refactor to go-kit style endpoints. Refs #1
This is a slight abuse of the framework, since go-kit is RPC and
fundamentally message-oriented. Still, it's worth seeing if we can
wrap endpoints in middleware without loading the entire request
object into memory.

The benefit of using go-kit middleware will be adopting a well-
defined pattern for logging, backpressure, service discovery, etc.

If it doesn't work for the /upload endpoint, we can implement that
with straight-up "net/http" and only use go-kit for RPC-style
operations.