Hacking some of the generated thrift code.
The reason I'm hacking generated code is that some of the fields are
marked 'required' in the generated code, but the AAC does not return
them.

This is code in the non-Apache library. We are using a non-Apache
library because it is easy to create a Thrift client with just a
standard, tls-enabled net.Conn object. It is this object that the third
party Golang OpenSSL wrapper library returns to us.