Edits to generated code to support BuildAcm() method. Refs #2.
Added a utils package to convert strings to []int8.

The required edits to generated code to support the BuildAcm()
method are simply replacing every instance of `[]byte` with `[]int8`
in the generated aacservice.go file.

Calls into BuildAcm() require an xml snippet, and this is expected
to be a string that requires conversion to `[]int8`, which is the
Golang equivalent of a Java byte array. AAC is implemented in Java
and Scala.