Serve static files from dedicated /static dir.
Added a StaticPath field to AppServer to be set on startup. Any
JavaScript or CSS files should be added to their respective
subfolders.

A single handler now can grab the filepath following the StaticDir
path with a regex utility function that gets a map of all named
variable capture groups.

Relative paths, if requested, are sanitized. See the file regex_test.go
for an example of the type of paths that are sanitized. We return a 404
for these paths.

Refs #54