Cannot undelete expunged objects.
The DAO gets an object by ID, and the controller checks the IsExpunged
field. If true, return error code 410.

The nil check on the stats channel in sendErrorResponse is required, due
to the fact that the test runs against the fake server. A send to a nil
channel blocks forever. See:
http://dave.cheney.net/2014/03/19/channel-axioms

Also removed punctuation from some error messages, as this is against
the Go style guide.

Fixes #125