[WIP] make sure that db record is written before S3 drain
gracefully handle the case where one client starts caching a file,
and another also wants to download it.  one goroutine is in charge of
caching a file, and the others need to wait for that to finish.

we also handle the case where a client asks for a file, then
cancels the request because it is not cached (taking a long time
to even begin downloading).  caching the file will continue
even if the user disconnects, so that upon return later, it
will be in the cache.

this is preparation for handling the third case, where we
cannot yet start caching it because it is still uploading
into S3 in from a different metadataconnector instance
(sharing the same database and bucket).  we will only fully
exercise this case with multiple uploaders against the same database.

include inline disposition

add in a unit test