New CLI interface. New config odrive.yml
Added test command to test AWS S3 bucket permissions. You need to set this
in your environment:
   export OD_AWS_S3_BUCKET=decipherers

Environmnet variables are now string constants in their own source code
file. If you add a new variable, please add it to this file, and update
the function that prints the current environment to stdout.

Deprecate conf.json in favor of odrive.yml. This new config file only
contains the ACL whitelist so far. Other structures can be added to the
mapping ConfigFile struct, but we should minimize the need for files.

Default constructor for AppConfiguration was added for tests, because
old tests relied on their own conf.json in the dao package.