Script for building an rpm.
Expects ODRIVE_BINARY_DIR and ODRIVE_VERSION to be set. We can do this with
Salt, or at the command line. ODRIVE_BINARY_DIR corresponds to
**cmd/odrive**. This directory is expected to contain a
centos-compatible binary as well as odrive.yml

If you are on a centos machine with rpm-build installed, this script
should produce an RPM that wraps a binary and the basic odrive.yml
config. Binary will be unpacked into /usr/bin and config will be
unpacked to /etc/odrive.

Also updated odutil utility program to allow specifying a "key" to write
to S3. Previously the key was the same as whatever file was passed to
`input` flag. Now that we can write a custom key, we can have nested
paths like:

```
odrive-builds/rpms/odrive-1.0.x86_64.rpm
```

Also add s3 library to govendor.