Pass down CommandLineOpts object to config constructors.
All CLI options should be set on CommandLineOpts and passed down through
the `startApplication(...)` function and on to the main AppConfiguration
constructors. This makes the code cleaner and further centralizes
config.

Reading in the yaml file is still separate, and we are passing down the
whitelist `[]string` on its own still. This can be refactored later.

Refs #69