Refactor test code
Create helper functions for permissions in server_test. These rely on
the order of the package global `clients` of type []*ClientIdentity,
constucted in server_setup_test.go

Remove global httpclients, and change tests to use to clients[idx].Client
because this field contains the TLS-enabled HTTP client associated with
the test client index. Client is a new field.

Refs #108