Fix for #117: Support for listing objects shared to everyone. Fix for #119: composite permissions
fix sql error in objectacm trigger

Add table driven test for expected counts of db schema components. This
is a sanity check for the future. Other cases can be added to the anonymous
[]struct{}

Refs #122

Fix for #122 - Trigger updated

Fix for #124 - Check if existing normalized ACM has really changed before updating.

There are scenarios where application code tries to update the acm
associated with an object, but the corresponding acm table entry
will not be different. This can cause errors at the database level,
because the triggers that fire in response to UPDATE statements
expect values to have actually changed.

Refs #122

Resolving nested struct