Async integration tests for event publisher
Many API calls have emitting Kafka events as a side effect, and these
tests ensure that we are publishing events. Here we test that a create,
update, and delete event are published to Kafka during our test.

As with most integration tests, this requires adjustment to CI configs.

Also remove dead code in testhelpers.

Refs #29