Return an array of breadcrumb objects on getObject calls. #96
Declare a protocol type Breadcrumb, and embed []protocol.Breadcrumb in
protocol.Object responses for explicit "get object" API calls, e.g.

> /objects/<objectID>/properties

Sorting for breacrumbs is determined by walking up the linked-list of
parentIDs via a new DAO method, GetParents.

Clients that want to display a list of objects under a folder as well as
the breadcrumbs for that folder will need to do two API calls: one to
list the objects and another to get the properties for the folder, which
will include breadcrumbs.

Breadcrumbs will most often represent folders, and the folder names
associated with Breadcrumbs will be redacted to the string "Not
Authorized" if 1) the user does not have clearance, or 2) the objects
are not shared to the user.