Add a test to assert the length of our big list of config options.
Every odrive config option is available as an environment variable
declared as a const in envvars.go. We also add those const identifiers
to a slice, which is useful for iterating/printing the config vars that
can be set.

This change adds a test that makes sure the length of the Vars slice is
the same as the number of consts declared in the file envvars.go and the
Golang ast parser is required for this task. Supports #152