Filter an object's breadcrumbs if any parent is inaccessible
DAO method GetParents still does no filtering, and returns the
entire parent as a root-first-sorted slice. Filtering is done
in the redactParents function.

Only the "closest" parents are returned in breadcrumbs. After an
inaccessible parent is encounted, we stop walking up to root and return
that slice of accessible parents.

Remove checks for redacted names in breadcrumb tests because redacted
breadcrumbs are no longer returned.

Fixes #211