Treat ownedby field as serialized resource string
This is a precursor to supporting changeOwner api operation and group
ownership.  Since operations that allow changing permissions and in
particular ownership need to ensure that the owner maintains full CRUDS
and rebuilding the ACM share requires all the parts for groups, adding
support for ownership to be by a group must be done such that the
ownership conveys all fields about the group.  These fields include the
project name, project display name, and finally the group name.

Recent enhancements have standardized on a serialized resource string
format for giving permissions in the create object and update object
calls, and so it is natural that the same structure be used when
indicating the owner of an object when changing from the default. The
default owner of an object when created is the user that created the
object.

* Existing objects are owned by users as `user/{distinguishedName}`
* odrive-database tool has migrate up/down script for transformation
* odrive-database tool now supports a rudimentary list command
* API Documentation updates for response object format
* Added test for ODAcmGrantee parsing of resource strings

Closes #22