Allow custom ZK to discover AAC.
If ZK addrs are provided for AAC, connect to that new ZK and track AAC
from it. We will log a message if a "custom" ZK is provided for AAC.

Also remove old service tracking code. We did not go with this pattern
in servicestates.go, so remove the dead code.

Refs #56