Provide ability to discover Kafka queue from ZK. Fixes #56
Introduces another way to configure the kafka-backed Publisher interface.

Two config options control the behavior of this feature:
OD_EVENT_ZK_ADDRS and OD_EVENT_KAFKA_ADDRS. If neither are set, a null
implementation is used. If OD_EVENT_KAFKA_ADDRS is set, it will take
precedence, and a direct connection will be attempted. If
OD_EVENT_ZK_ADDRS is set, attempt to discover Kafka from ZK.

Set one or the other, not both. No default ZK will be used as a
fallback, so one of the configs MUST be provided to publish to Kafka.

Other changes:
Rename package finder to package kafka. Test for DiscoverKafka func.