We should not allow gatekeeper to send us work until we are ready
Delay announcement to zookeeper until all known prerequisites for correct responses are ready.
So far, it seems like an AAC check and the http Serve function running are sufficient.
So, we sleep for 1 second intervals while the app.AAC is still nil.

If there are multiple odrives running, and at least one is running, we should now not get transient errors
if a client is sent to an odrive that just recently started.  Previously, if running integration tests
before "aac chosen" happened, the tests could fail, even if odrives were already mapped by gatekeeper.

Fixes #293