Retry getting and creating users in database to mitigate race conditions
On each request, we look up the caller in the user cache, and if that
cache misses, we go to the database. If many requests from a new user
happen at once, multiple goroutines can attempt to create a user at the
same time. This will be prevented by the database, but it is a
recoverable error.

Attempt to recover from this error by simply trying to get or create the
user twice. If another goroutine creates the user first, the subsequent
call to dao.GetUserByDistinguishedName should succeed.

Fixes #283