add cors support
User interfaces that are not mounted into gatekeeper would like to use odrive.
The cross-site requests require us to put in a value for Access-Control-Allow-Origin header.
The most secure thing (for the UI) is for use to reflect back the value of Origin, which must be set
on requests that are dealing with cross-site requests.

The unit tests verify that we reflect back Origin as Access-Control-Allow-Origin on all requests.
It also ensures that the OPTIONS method reflects back

* Access-Control-Request-Method - as something that includes this value in Access-Control-Allow-Method
* Access-Control-Request-Headers - as Access-Control-Allow-Headers

This was tested by having our backend hardcode the full url of the backend using a hostname,
and then accessing the site through a different hostname (that will map to same IP).  As far
as the browser is concerned, this is cross-site.  To test that way, go into listObject.js and temporarily
fully qualify BASE_SERVICE_URL, and access the UI via IP address or an alias for the IP that
does not match BASE_SERVICE_URL.

Fixes #306