run tests correctly from outside container or inside container
When running tests, be specific about which hostname to use, and let the DNS of the outside (/etc/hosts) or internal to container (127.0.0.11) correctly map them.
This should prevent the need to use docker from inside a container for cases where we need to be able to reach all machines in the network.

```
go test ./...
```

```
docker exec -ti docker_odrive_1 go test ../../...
```

This also requires that OD_DOCKERVM_OVERRIDE is passed in to the cte/object-drive project to actually use.
Make sure that when running tests outside the container, the hostnames are mapped.

Fixes #330

make circle happy
update readme
simplify picking zk connection