Base Centos low-side AMI build.
If you have `packer` installed, and provide the required assets to the
build (see README.md), this should be able to create an AMI that will
automatically start an object-drive-1.0 service.

Note that the exact configurations must be provided. Also note that in
high-side environments, the "base image" must be modified, which will
probably require modifications to the build. Packer relies on SSH to do
its work, so paths and users will need to be adjusted from the low-side
defaults.

Other changes:
* some code style cleanup