fix race condition on test after build
in addition to odrives only registering with zookeeper when they can serve requests without errors (important for scaling without transient errors),
we should be able to run the test suite immediately after a docker-compose up such that we know that at least gatekeeper and zookeeper are up.
this will keep us from having an arbitrary long sleep that guarantees that we will be ready to run tests.

a new /ping endpoint was added, that does nothing at all except return a 200, to deal with the fact that the nginx server will return
error codes of its own when it cannot connect to an instance of odrive in the backend.

respond to comments, and use channel to monitor startup wait
ensure that we stop the ticker

Fixes #294