empty trash operation
operation to iterate trashed objects and delete them

`DELETE /trashed` will delete a large number of objects (10000 default).
In order to allow the time to be bounded os that trash cleaning can stay bounded,
it is also acceptable to do `DELETE /trashed?pageSize=100`.

The transaction will try to delete all trash.  But the transaction will be segmented into chunks of size pageSize,
so that it is guaranteed that we make timely progress; and cannot get stuck because http timeouts are longer
than time to take out all trash.

```
  {"expunged_count": 35}
```

Fixes #348

back out change to wrong branch