ensure that static file requests dont escape the root
there are many combinations of patterns that can be used to defeat checks
that we are not trying to escape the tree.  what is important is that
at the end of normalization with path.Clean, we have a file that is definitely
somewhere under h.StaticDir, and that any ".." have been flattened out of the path.

It is crucial that we positively check that we are still under h.StaticDir.

Also, because these are our own static files that we control the names of,
just to remove all doubts about safety, just completely allow
all escaping of URLs.  We cannot control if the original URL needs to use
relative pathing (it's a requirement that we allow it in order to function
normally with a browser).  We simply need to ensure that after the path
is clean that we are still under h.StaticDir.

Fixes #472