we are getting failures due to default certs expiring. this is enough to fix that problem, and there is no intent to clean up messes within this MR.
But note:

It appears that we may be able to get rid of cgo.  But leave that for another commit to prevent us from debugging multiple issues at once.

We need to override certs in a repo that we are not checking out here:

$OD_ROOT/dias-simulator/nginx/ssl/{ca.crt,server.crt,server.key}

This build will fail in circle until these containers are pushed:

  gatekeeper
  odrive
  packager
  dias-simulator