Improve durability of AAC connection
Reduces unnecessary RPC client shutdown on snippet error response not setting snippets
Added test for retrieving objects for a non-existent user
Updated tests that leverage the server cert as a user to always expect failure.
Depends on alterations to DIAS simulator that does not return default profile for non-existent user.

Closes #523