fault barrier in purge thread
every goroutine in which a panic happens has the potential to bring down the server.
this fault barrier was not added previously, just because we never got a crash here.
but we should catch here, as we do in the top of ServeHttp, because we do not
use the strategy of crashing and simply waiting for something else to respawn us.

Fixes #556

respond to comments

respond to comments