Set versions at build time inside rpm container
CircleCI runs our packaging step inside of a container. This runs an
additional go build, so we have to pass -ldflags during this additional
build.

Build metadata and git tag names must be passed down:

    circle.yml -> makerpm -> build.sh

The final packaging step in prepare-rpm-env.sh does not do a go build,
so we do not need to pass ldflags there. We might consider doing this
later, to embed a more accurate version in the rpm itself. Currently the
rpm's Version field is the CircleCI build number.

Fixes #600