Fixed fatal on service start
S3 read/writes can throw error if to many requests occur in a short window of time. If this happens during service startup, it causes a fatal crash. A random retry has been put in to help accomodate this edge case, as well as improved logging when using S3 vs when not. (949)
Minimal set of docker configuration prepared and noted in the readme. Some further cleanup done to readme. (1101)
Removed dead code in crypto module
Some efforts to normalize on the use of the term 'permanent storage' as opposed to 'persistent storage'

Fixes #949
Fixes #1101