Don't default OD_PEER_CN
Documentation denoted it is not defaulted. Implementation had it defaulted which is erroneous. Anytime there is a dependency on it due to multiple peers in an instance, the CN of the peer should be explicitly defined by setting OD_PEER_CN environment variable (This, along with OD_PEER_SIGNIFIER do not support configuration files with cascade support).

Also converted some more locations that should be referencing constants for the environment variable names.

Fixes #1085